pyproject.toml
[project]
name = "WrapConfig"
version = "0.2.1"
description = "Wraper to manage configurations"
authors = [{name = "Julian Kimmig", email = "julian.kimmig@linkdlab.de"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
toml = [
    "toml>=0.10.2",
]
yaml = [
    "pyyaml>=6.0.2",
]
all = [
    "toml>=0.10.2",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "coverage>=7.6.10",
    "pytest>=8.3.4",
    "WrapConfig[toml]",
    "WrapConfig[yaml]",
]
toml = [
    "WrapConfig[toml]",
]
yaml = [
    "WrapConfig[yaml]",
]
all = [
    "WrapConfig[toml]",
    "WrapConfig[yaml]",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[project.license]
text = "MIT"

[project.urls]
homepage  = "https://github.com/JulianKimmig/WrapConfig"
source  = "https://github.com/JulianKimmig/WrapConfig"
tracker  = "https://github.com/JulianKimmig/WrapConfig/issues"

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"





